map.scale {maps}	R Documentation

Add Scale to Existing Unprojected Map

Description

Adds a scale to an existing map, both as a ratio and a distance gauge.

Usage

map.scale(x, y, relwidth = 0.15, metric = TRUE, ratio = TRUE, ...)

Arguments

x

y
Location of left end of distance gauge. If not specified, this will be taken to be near the lower left corner of the map.

relwidth
Proportion of width of display to be used for the scale. The default is 0.15 (15%).

metric
If TRUE, the distance gauge will be in km, otherwise miles.

ratio
If FALSE, the scale ratio of the map is not displayed.

...
Further plotting parameters may be specified as for the command text().

Details

The scale is calculated from the displayed graph's plotting parameters, and the latitude of the location at which the distance gauge will be displayed.

Value

The exact calculated scale is returned.

NOTE

This function is meaningful only if no projection= has been specified in the call to map().

Side Effects

A scale is added to the currently displayed map. This takes the form of an approximate 1:n scale (containing 2-3 significant digits), above a distance gauge which is reasonably accurate for the latitude at which it appears. The circumference at the given latitude is interpolated from a radius of 6356.78 km at the pole and 6378.16 km at the equator.

See Also

map.axes

Examples

map("world", "China")
map.scale()

[Package maps version 3.3.0 Index]