ozone {maps}	R Documentation

Sample datasets

Description

Datasets used to illustrate map functions.

ozone contains the median of daily maxima ozone concentration in 41 US cities for June 1974 through August 1974. Concentrations are in parts per billion (ppb).

unemp Has population and unemployment percentage for US counties.

votes.repub contains the percentage republican votes in the 1900 election.

Usage

data(ozone)
data(unemp)
data(votes.repub)

References

Cleveland, W.S., Kleiner, B., McRae, J.E., Warner, J.L., and Pasceri, P.E. , "The Analysis of Ground-Level Ozone Data from New Jersey, New York, Connecticut, and Massachusetts: Data Quality Assessment and Temporal and Geographical Properties", Bell Laboratories Memorandum, 1975.

[Package maps version 3.3.0 Index]