state.carto {maps}	R Documentation

United States State Population Cartogram Map

Description

This database produces a cartogram of the states of the United States mainland based on CartoDraw, roughly proportional to population (see references).

state.carto.center are coordinates of the state centers for annotation purposes.

Usage

data(stateMapEnv)
data(state.carto.center)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR for the datasets in the maps package) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

CartoDraw, http://www.computer.org/csdl/trans/tg/2004/01/v0095-abs.html

See Also

map.

Examples

map('state.carto', fill = TRUE, col = palette())

[Package maps version 3.3.0 Index]