state.fips {maps}	R Documentation

FIPS state codes for US 48 State Map

Description

A database matching FIPS codes to maps package state names.

Usage

data(state.fips)

Format

A list with 6 components, namely "fips", "ssa", "region", "division", "abb" and "polyname", containing the US Census Bureau FIPS, SSA, REGION and DIVISION numbers, the standard state abbreviation and the respective state polygon name. Note that "fips" is represented as an integer, so any leading zero (which is part of the fips code) is not shown by default.

See Also

county.fips

[Package maps version 3.3.0 Index]