state.vbm {maps}	R Documentation

United States State Visibility Base Map

Description

This database produces a map of the states of the United States mainland. The Visibility Base Map was created by Mark Monmonier to provide simplified state shapes with sufficient areas to allow annotations in even the small states.

state.vbm.center are coordinates of the state centers for annotation purposes. The states are alphabetically ordered, in the same order as the map. So state names can be matched via e.g. map(state.vbm, plot=FALSE)$name.

Usage

data(state.vbmMapEnv)
data(state.vbm.center)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR for the datasets in the maps package) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

Mark Monmonier and George Schnell, "The Study of Population", Elements, Patterns, Processes. Charles E. Merrill. Columbus, OH. 1982.

See Also

map.

Examples

map('state.vbm', fill = TRUE, col = palette())

[Package maps version 3.3.0 Index]