world {maps}	R Documentation

Low (mid) resolution World Map

Description

This world map (updated in 2013) is imported from the public domain Natural Earth project (the 1:50m resolution version). It replaces a much older version based on the CIA World Data Bank II data. The old legacy data is still available in the package mapdata (v2.3.0).

Usage

data(worldMapEnv)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR_WORLD) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

Details

As of version 3.1, the world database no longer contains any lakes. These have been moved to a separate database called lakes. The legacy world map (dating from around 1990) has been removed from the package and is now available from the mapdata package in two different resolutions (worldHires and worldLores).

Source

The Natural Earth data set is in the public domain and available from https://www.naturalearthdata.com.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

See Also

map,lakes

Examples

# notice how some polygons extend beyond the [-180,180] interval:
map('world', fill = TRUE, col = 1:10)
# if you wrap at [-180,180], you also can get a clean closure of Antarctica
map('world', fill = TRUE, col = 1:10, wrap=c(-180,180) )

[Package maps version 3.4.0 Index]