world.cities {maps}	R Documentation

Database of world cities

Description

This database is primarily of world cities of population greater than about 40,000. Also included are capital cities of any population size, and many smaller towns.

Usage

data(world.cities)

Format

A list with 6 components, namely "name", "country.etc", "pop", "lat", "long", and "capital", containing the city name, the country name, approximate population (as at January 2006), latitude, longitude and capital status indication (0 for non-capital, 1 for capital, 2 for China Municipalities, and 3 for China Provincial capitals)

NOTE

Some of the country names and city names may be out of date. Please send any corrections to the package maintainer.

Source

The data were originally obtained from Stefan Helders' website http://www.world-gazetteer.com, which now redirects to http://www.populationmondiale.com. There are no recent updates available.

See Also

map.cities

[Package maps version 3.3.0 Index]