world2 {maps}	R Documentation

Pacific Centric Low resolution World Map

Description

This is an alternative version of the world database based on latitudes [0, 360), which then has the Pacific Ocean in the centre of the map.

Usage

data(world2MapEnv)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR_WORLD for the datasets in the maps package) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

NOTE

This data set is in fact largely obsolete. Often the same (more general) result can be obtained by using wrapping:

map("world", wrap=c(0,360))

This will also work fine with fill=TRUE or any other appropriate longitude interval (e.g. c(-90,270)).

However, world2 is useful when setting xlim to an interval crossing the 180 meridian.

Source

The Natural Earth data set is available from http://www.naturalearthdata.com.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

See Also

map, world

Examples

map('world2', xlim = c(100, 300))
map.axes()
# xlim is performed before wrapping:
map('world', wrap=c(0,360), xlim = c(100, 300))
# so to emulate "world2":
ww2 <- map('world', wrap=c(0,360), plot=FALSE, fill=TRUE)
map(ww2, xlim = c(100, 300), fill=TRUE)

[Package maps version 3.3.0 Index]