canada.cities {maps}	R Documentation

Database of Canadian cities

Description

This database is of Canadian cities of population greater than about 1,000. Also included are province capitals of any population size.

Format

A list with 6 components, namely "name", "country.etc", "pop", "lat", "long", and "capital", containing the city name, the province abbreviation, approximate population (as at January 2006), latitude, longitude and capital status indication (0 for non-capital, 1 for capital, 2 for provincial

capital.

NOTE

Some of the city names may be out of date. Please send any corrections to the package maintainer.

See Also

map.cities

[Package maps version 3.1.0 Index]