county.fips {maps}	R Documentation

FIPS county codes for US County Map

Description

A database matching FIPS codes to maps package county and state names.

Usage

data(county.fips)

Format

A list with 2 components, namely "fips" and "polyname", containing the FIPS number and respective state or county polygon name. Note that "fips" is represented as an integer, so any leading zero (which is part of the fips code) is not shown by default.

See Also

state.fips

[Package maps version 3.3.0 Index]