county.fips {maps}	R Documentation

FIPS county codes for US County Map

Description

A database matching FIPS codes to maps package county and state names.

Usage

data(county.fips)

Format

A list with 2 components, namely "fips" and "polyname", containing the FIPS number and respective state or county polygon name.

See Also

state.fips

[Package maps version 3.1.0 Index]