legacy_world {maps}	R Documentation

Legacy low resolution World Map.

Description

This world database comes from a thinned cleaned-up version of the CIA World Data Bank II data and contains approximately 30,000 points representing the world coastlines and national boundaries approximatly around 1990.

Usage

data(worldMapEnv)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR for the datasets in the maps package) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

During a transition period (maps version 3.0) it is possible to switch to the old legacy map using the function world.legacy(TRUE) or by setting the setting the environment variable R_MAP_DATA_LEGACY=TRUE prior to loading the package.

Source

The CIA World Data Bank II is currently (2015) still available from https://www.evl.uic.edu/pape/data/WDB/.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

See Also

map,world.legacy,world

Examples

map('legacy_world', fill = TRUE, col = 1:10)

[Package maps version 3.1.0 Index]