world {maps}	R Documentation

Low (mid) resolution World Map

Description

This updated (2013) world map is imported from the Natural Earth project (the 1:50m resolution version). It replaces a much older version based on the CIA World Data Bank II data. That older version is still included in the package and can reached as legacy_world.

Usage

data(worldMapEnv)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR_WORLD) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

During a transition period (maps version 3.0–1) it is possible to switch to the old legacy map using the function world.legacy(TRUE) or by setting the setting the environment variable R_MAP_DATA_LEGACY=TRUE prior to loading the package.

Details

As of version 3.1, the world database no longer contains any lakes. These have been moved to a separate database called lakes.

Source

The Natural Earth data set is available from http://www.naturalearthdata.com.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

See Also

map,world.legacy,legacy_world,lakes

Examples

map('world', fill = TRUE, col = 1:10)

[Package maps version 3.1.0 Index]