world2 {maps}	R Documentation

Pacific Centric Low resolution World Map

Description

This is an alternative version of the world database based on latitudes [0, 360), which then has the Pacific Ocean in the centre of the map.

Usage

data(world2MapEnv)

Format

The data file is merely a character string which specifies the name of an environment variable which contains the base location of the binary files used by the map drawing functions. This environment variable (R_MAP_DATA_DIR_WORLD for the datasets in the maps package) is set at package load time if it does not already exist. Hence setting the environment variable before loading the package can override the default location of the binary datasets.

During a transition period (maps version 3.0–1) it is possible to switch to the old legacy map using the function world.legacy(TRUE) or by setting the setting the environment variable R_MAP_DATA_LEGACY=TRUE prior to loading the package.

NOTE

Almost the same effect can in fact be obtained by using a projection:

map("world",projection="rectangular",parameter=0,orientation=c(90,0,180), wrap=TRUE)

By changing the third component of 'orientation', the longitude shift can also be set to any other value besides 180 or 0 (but then fill=TRUE will usually give artefacts at the borders). You may even need to set 'resolution=0' if the wrapping algorithm fails because of the thinning.

This projected version will not work so well with fill=TRUE, though. There may be jagged edges at the boundaries.

Source

The Natural Earth data set is available from http://www.naturalearthdata.com.

References

Richard A. Becker, and Allan R. Wilks, "Maps in S", AT\&T Bell Laboratories Statistics Research Report [93.2], 1993.

Richard A. Becker, and Allan R. Wilks, "Constructing a Geographical Database", AT\&T Bell Laboratories Statistics Research Report [95.2], 1995.

See Also

map, world,legacy_world2

Examples

map('world2', xlim = c(100, 300))
map.axes()

[Package maps version 3.1.0 Index]